created rel-word-position

+++ lib/boot.rb
@@ -7,47 +7,47 @@ DB = Sequel.sqlite
 DB.create_table :words do
   primary_key :id
   String :text
-  Integer :left
-  Integer :right
-  Integer :top
-  Integer :bottom
+  Float :left
+  Float :right
+  Float :top
+  Float :bottom
 end
 
 DB.create_table :price_terms do
   primary_key :id
   String :text
-  Integer :left
-  Integer :right
-  Integer :top
-  Integer :bottom
+  Float :left
+  Float :right
+  Float :top
+  Float :bottom
 end
 
 DB.create_table :date_terms do
   primary_key :id
   String :text
-  Integer :left
-  Integer :right
-  Integer :top
-  Integer :bottom
-  Integer :first_word_id
+  Float :left
+  Float :right
+  Float :top
+  Float :bottom
+  Float :first_word_id
 end
 
 DB.create_table :vat_number_terms do
   primary_key :id
   String :text
-  Integer :left
-  Integer :right
-  Integer :top
-  Integer :bottom
+  Float :left
+  Float :right
+  Float :top
+  Float :bottom
 end
 
 DB.create_table :iban_terms do
   primary_key :id
-  String  :text
-  Integer :left
-  Integer :right
-  Integer :top
-  Integer :bottom
+  String :text
+  Float :left
+  Float :right
+  Float :top
+  Float :bottom
 end
 
 DB.create_table :billing_period_terms do
@@ -59,10 +59,10 @@ end
 DB.create_table :currency_terms do
   primary_key :id
   String :text
-  Integer :left
-  Integer :right
-  Integer :top
-  Integer :bottom
+  Float :left
+  Float :right
+  Float :top
+  Float :bottom
 end
 
 Sequel::Model.db = DB
